backend/requirements/dev.txt
# Development Requirements
-r base.txt

# Testing
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0

# Code quality
black>=24.0,<25.0
ruff>=0.1,<1.0
mypy>=1.8,<2.0

# Development tools
ipython>=8.0,<9.0
django-extensions>=3.2,<4.0
